Fixed initialization of ECS. Added simple text debugging. Broke most other examples by moving implementation of efi_global_state to a c file.

+++ examples/ecs/ecs.h
@@ -6,8 +6,8 @@
 #include "entity.h"
 #include "system.h"
 
-static const int PAGE_SIZE = 4096; // Assuming a page size of 4096 bytes (what we get from the UEFI allocator))
-static const int POOL_SIZE = 4096; // Number of instances of entities / instances per component type to allocate for
+static const uint64_t PAGE_SIZE = 4096; // Assuming a page size of 4096 bytes (what we get from the UEFI allocator))
+static const uint64_t POOL_SIZE = 4096; // Number of instances of entities / instances per component type to allocate for
 
 typedef struct ECS {
     EntityData entityData; // Array of all entities
